+++ src/App.tsx
@@ -7,6 +7,8 @@ import { HomePage } from "./pages/HomePage";
 import { HistoryPage } from "./pages/HistoryPage";
 import { ChannelsPage } from "./pages/ChannelsPage";
 import { ChannelPage } from "./pages/ChannelPage";
+import { PlaylistsPage } from "./pages/PlaylistsPage";
+import { PlaylistPage } from "./pages/PlaylistPage";
 import { WatchPage } from "./pages/WatchPage";
 import { LibraryPage } from "./pages/LibraryPage";
 
@@ -32,8 +34,12 @@ function AppInner() {
         <Route element={<Shell />}>
           <Route path="/" element={<HomePage />} />
           <Route path="/history" element={<HistoryPage />} />
-          <Route path="/channels" element={<ChannelsPage />} />
+          <Route path="/folders" element={<ChannelsPage />} />
+          <Route path="/folder/:slug" element={<ChannelPage />} />
+          <Route path="/channels" element={<Navigate to="/folders" replace />} />
           <Route path="/channel/:slug" element={<ChannelPage />} />
+          <Route path="/playlists" element={<PlaylistsPage />} />
+          <Route path="/playlist/:id" element={<PlaylistPage />} />
           <Route path="/watch/:id" element={<WatchPage />} />
           <Route path="/library" element={<LibraryPage />} />
           <Route path="*" element={<Navigate to="/" replace />} />